refactor: consolidate tests into existing test files
- Move collection tests to test_collection.py
- Move file resource tests to test_data.py
- Move privilege v2 tests to test_user_role.py
- Remove temporary test files
- Full test suite: 91 passed, 2 skipped
